+++ CHANGES.rst
@@ -16,6 +16,11 @@ Bug Fixes
 
 - 
 
+ramp_fitting
+~~~~~~~~~~~~
+
+- Fixed memory leak in C-extension.[#281]
+
 1.8.0 (2024-08-14)
 ==================